+++ src/db/schemas/Card.js
@@ -1,9 +1,12 @@
 const { Schema } = require("mongoose");
-const PlayerSchema = require("./Player");
-const { getCardCategories } = require("../../helpers/functions/Card");
+const { getCardCategories, getCardStatuses } = require("../../helpers/functions/Card");
 
 const CardSchema = new Schema({
     label: { type: String },
+    status: {
+        type: String,
+        enum: getCardStatuses(),
+    },
     categories: {
         type: [String],
         enum: getCardCategories(),
@@ -13,13 +16,6 @@ const CardSchema = new Schema({
         type: Number,
         min: 1,
         max: 3,
-    },
-    guessed: {
-        from: {
-            type: PlayerSchema,
-            required: true,
-        },
-        by: { type: PlayerSchema },
     },
     description: { type: String },
     imageURL: { type: String },