+++ src/helpers/functions/Card.js
@@ -1,3 +1,5 @@
-const { cardCategories } = require("../constants/Card");
+const { cardCategories, cardStatuses } = require("../constants/Card");
 
 exports.getCardCategories = () => JSON.parse(JSON.stringify(cardCategories));
+
+exports.getCardStatuses = () => JSON.parse(JSON.stringify(cardStatuses));